Fixup missing/empty relationships
If include_data=False we must return meta or links. This caused a change
to a test which predates this commit that had a related url which is
does not conform to jsonapi standard. I.e. the id of the related
resource can not be relied upon.

And now the related *resource linkage*, i.e. 'data' attribute of the
related resource object resolves to null or []

Also added test for include_data=False