Add support for automatic self-linking
This works through the new Meta options `self_url`, `self_url_kwargs`
and `self_url_many`. Specifying these will add a link object with a self
link to the top-level and each resource.

Tests and documentation updates are included, as are a new
AUTHORS.rst-file to comply the with contribution guidelines.