Fix included_data and data["included"] tpe consistency.
included resource objects in the json api is a list but the schema implementation
needs it to be a dict indexed by type and id.
The new test checks that loading and then dumping with the same schema object on a
dict with an included resource works. It is needed for Flask-rest-jsonapi that re-uses schema between load and dump.